+++ crates/wasmtime/src/runtime.rs
@@ -1,3 +1,5 @@
+#![cfg_attr(nightlydoc, doc(cfg(feature = "runtime")))]
+
 #[macro_use]
 pub(crate) mod func;
